// dashboard/src/api/models/user.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export interface User {
  id: number;
  readonly name: string;
  email: string;
  /** @nullable */
  avatar_url: string | null;
  avatar_id: string;
  readonly active_account_id: string;
  joined_at: string;
  readonly has_usable_password: boolean;
}